requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2023.1
Django==5.0.1
django-browser-reload==1.12.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-tailwind==3.8.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
fastapi==0.109.2
h11==0.14.0
idna==3.6
joblib==1.3.2
mod_wsgi==5.0.0
multipart==0.2.4
numpy==1.26.3
openpyxl==3.1.2
pandas==2.2.0
pillow==10.2.0
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.8.2
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2024.1
reportlab==4.0.9
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.36.3
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
uv==0.1.6
uvicorn==0.27.1
